Separated Navbar and Navheader components from page contents.
